// src/string-extract/UserConfiguration.d.ts
import type { NodeDef } from 'node-red';
import { MappedOutput } from './MappedOutput';

/**
 * Configuration options available to users.
 */
export interface UserConfigurationOptions {
    extractionMode?: string;
    inputValueProperty?: string;
    inputValueSource?: string;
    extractionRegexpPattern?: string;
    extractionRegexpFlags?: string[];
    extractionSplitSeparator?: string;
    outputMapping: Array<MappedOutput>;
}

/**
 * Configuration generated by the user in the editor.
 */
export interface UserConfiguration extends NodeDef, UserConfigurationOptions {}